Add seeding function for database
In order for the application to function properly, it first needs to be
seeded with all the companies. This commit adds a seeding function at
Nexus.Seeds that reads the data from the CSV file attached and inserts
it into a sqlite3 database.

Using the list of companies provided at seeds.csv, the seeding function
attempts to reach the Lever API to get a JSON response of all current
openings for each company. If the API call is successful, the response
body gets hashed and stored alongside the company name. This allows
for a simple check to occur at any given moment to see if the data
has changed in any way (if a job has been added or removed).

The seeding function is called from the seeds.exs file, which needs
to be ran before the application is started when running in :prod.

In order to seed the database, two new dependencies were added:
    - csv -> 3.2
    - HTTPoison -> 2.0