Add support for elasticsearch
On top of using a normal sqlite3 database to handle the data, the more
complex queries will be handled by elasticsearch. This is because the
jobs have more tags that benefit with the introduction of a search
index.

The elasticsearch server is expected to be running on localhost:9200
when in :dev. A ca.crt file is expected to reside in 'priv/certs' in
order to connect using 'https' which is the default.

The implementation only goes as far as a single function called put/1,
as nothing more than the seeding functions have been created. This acts
as the 'seeding' mechanism for the elasticsearch index.