Fixed elasticsearch functionality
The intital addition of elasticsearch happened to be premature as
it failed to account for a few things, all of which are now fixed.

The first is the basic authentication which is necessary to
connect to elasticsearch. This has now been added as an additional
environment variable titled:
	- ELASTICSEARCH_USERNAME
	- ELASTICSEARCH_PASSWORD

The second was fixing the way HTTPoison needed the ca.crt file to
connect to the elasticsearch API. It now connects and is able to
send PUT requests.