Clean json input for elasticsearch
A QoL fix, removing unnecessary fields, and adding in a hash function
to allow for quick comparison of the data, if case two jobs are the
same but have different IDs.

A docker-compose.yml file was also removed, as it was not needed for
the project. Although credentials were leaked, this server no longer
exists, and as such is not a problem.