Remove warnings within elasticsearch
When running the seeding program, the following warning was being
displayed:
    `warning: Application.get_env/2 is discouraged in the module body,
    use Application.compile_env/3 instead`

In order to follow better practice, its recommended to move all runtime
environment variables to be within a function. This commit removes all
the macros and instead places them where they are meant to be used (in
lieu of the macro).