Add daemon to update databases
In order to ensure that data is kept up to date, a few changes were
made to the application. The first change was to add a new module
`Nexus.Scheduler` which will be responsible for scheduling the update
using a new dependency `quantum`.

The second change was to the `Nexus.Elasticsearch` module to add a
simple check to see if the associated ID is already in the database.
If it is not, then to proceed with adding it. This allows for any new
jobs to be added to the database without having to worry about
conflicts.

To facilitate this action, a new module `Nexus.Update` was added to
house the internal logic of running through the sqlite3 database and
comparing if the cached_hash is the same as the current hash. Any
change in the hash will trigger an update to the elasticsearch database
as well as the sqlite3 database.