Decode json response during update cycle
The response from the Lever API is a JSON, and was being passed into
the `Nexus.Elasticsearch.put/1` straight as is.

This is incorrect, and using the working reference of seeds.exs,
the response needs to be decoded first and then passed into the function
to update the document.

A oversight on my part, can't believe I missed this.